Create new MeterBinder with Hibernate per Query Metrics
